Zephyr fix for newlib/minimallibc: to malloc.h or not to malloc.h?
When we originally made the port to Zephyr, if the Zephyr minimal C library was used [the default], instead of newlib, then malloc.h would not be included in the build, resulting in a failure to compile; since we need malloc()/free() (u_port_clib.c maps malloc()/free() to the Zephyr minimal C library functions) we needed a malloc.h from somewhere and so we put an empty one in the Zephyr source directory, letting the malloc.h from arm-none-eabi/include take over when newlib was used.

At some point this changed, or the include directory order changed, or maybe the file libc-hooks.c was introduced, not entirely sure, whatever: it is now true that our empty malloc.h prevails even in the newlib case, resulting in the compiler complaining that, when the internal Zephyr file libc-hooks.c is compiled, it has no prototype for malloc.h()/free() (which is true since it is picking up an empty malloc.h).

With this commit the empty malloc.h in the port/platform/zephyr/src directory is deleted and all is good with the world.